React on HTTP level errors in view request
Couchbase server DP4 allows unauthorized requests to views, but forbids ones
with valid Authorization header. It means that previous version of the ruby
client (more precisely underlying libcouchbase) works well with DP4 build
because it doesn't send Authorization header. But it will not work with
recent servers where the auth header is required for views.

Without this patch the client just hangs awaiting for completion and skips
the error.

Change-Id: I7d580002554b0950cee06e16b3eb11737514163a
Reviewed-on: http://review.couchbase.org/19436
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>