Broaden connection_pool version requirement
connection_pool 1.0 and 2.0 are API compatible and functionally
equivalent.  The major change for 2.0 is that the individual connections
are lazily created, instead of all at once.  This is particularly useful
when trying to use a pool of connections in something where the pool is
initialized before the Couchbase config (like using Sidekiq with
couchbase-model).

Change-Id: Ic72f5a0ac7eebf5b916917132f0ca1e0ade43944
Reviewed-on: http://review.couchbase.org/37439
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>