RCBC-451 & RCBC-452: Expose any specific lookup_in spec errors (#118)
Motivation
========
Any spec-specific errors were ignored which resulted in any error that happens being exposed PathNotFound (because of the absence of content)

Changes
=======
Add an error attribute to the SubDocumentField class that stores the Ruby error that is equivalent to the spec-specific error code from the C++ response. This error is raised then content or exists? for that spec are called

---------

Co-authored-by: Sergey Avseyev <sergey.avseyev@gmail.com>